split Picker into index and render
